requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
plotly>=5.14.0
streamlit>=1.22.0
duckdb>=0.9.0
pyyaml>=6.0
pytest>=7.0.0
ipykernel>=6.20.0
nbformat>=5.7.0
transformers>=4.0.0
torch>=2.0.0
